trekk ut runQuery
